+++ tests/Unit/Types/MiddlewareClassNamesTest.php
@@ -5,9 +5,7 @@ namespace IceHawk\IceHawk\Tests\Unit\Types;
 use IceHawk\IceHawk\Tests\Unit\Stubs\FallbackMiddleware;
 use IceHawk\IceHawk\Tests\Unit\Stubs\MiddlewareImplementation;
 use IceHawk\IceHawk\Tests\Unit\Stubs\PassThroughMiddleware;
-use IceHawk\IceHawk\Types\MiddlewareClassName;
 use IceHawk\IceHawk\Types\MiddlewareClassNames;
-use InvalidArgumentException;
 use PHPUnit\Framework\Exception;
 use PHPUnit\Framework\ExpectationFailedException;
 use PHPUnit\Framework\TestCase;
@@ -18,19 +16,18 @@ final class MiddlewareClassNamesTest extends TestCase
 	/**
 	 * @throws Exception
 	 * @throws ExpectationFailedException
-	 * @throws InvalidArgumentException
 	 */
 	public function testAdd() : void
 	{
 		$classNames = MiddlewareClassNames::new();
 		self::assertCount( 0, $classNames );
 
-		$classNames->add( MiddlewareClassName::new( MiddlewareImplementation::class ) );
+		$classNames->add( MiddlewareImplementation::class );
 		self::assertCount( 1, $classNames );
 
 		$classNames->add(
-			MiddlewareClassName::new( FallbackMiddleware::class ),
-			MiddlewareClassName::new( PassThroughMiddleware::class )
+			FallbackMiddleware::class,
+			PassThroughMiddleware::class
 		);
 		self::assertCount( 3, $classNames );
 	}
@@ -43,25 +40,7 @@ final class MiddlewareClassNamesTest extends TestCase
 	{
 		self::assertCount(
 			2,
-			MiddlewareClassNames::newFromStrings(
-				PassThroughMiddleware::class,
-				MiddlewareImplementation::class
-			)
-		);
-	}
-
-	/**
-	 * @throws ExpectationFailedException
-	 * @throws InvalidArgumentException
-	 */
-	public function testNewFromStrings() : void
-	{
-		self::assertEquals(
 			MiddlewareClassNames::new(
-				MiddlewareClassName::new( PassThroughMiddleware::class ),
-				MiddlewareClassName::new( MiddlewareImplementation::class )
-			),
-			MiddlewareClassNames::newFromStrings(
 				PassThroughMiddleware::class,
 				MiddlewareImplementation::class
 			)
@@ -71,46 +50,39 @@ final class MiddlewareClassNamesTest extends TestCase
 	/**
 	 * @throws Exception
 	 * @throws ExpectationFailedException
-	 * @throws InvalidArgumentException
 	 */
 	public function testNew() : void
 	{
 		self::assertCount( 0, MiddlewareClassNames::new() );
-		self::assertCount(
-			1,
-			MiddlewareClassNames::new(
-				MiddlewareClassName::new( MiddlewareImplementation::class )
-			)
-		);
+		self::assertCount( 1, MiddlewareClassNames::new( MiddlewareImplementation::class ) );
 		self::assertCount(
 			2,
 			MiddlewareClassNames::new(
-				MiddlewareClassName::new( PassThroughMiddleware::class ),
-				MiddlewareClassName::new( MiddlewareImplementation::class )
+				PassThroughMiddleware::class,
+				MiddlewareImplementation::class
 			)
 		);
 	}
 
 	/**
 	 * @throws ExpectationFailedException
-	 * @throws InvalidArgumentException
 	 */
 	public function testGetIterator() : void
 	{
-		$classNames = MiddlewareClassNames::newFromStrings(
+		$classNames = MiddlewareClassNames::new(
 			PassThroughMiddleware::class,
 			MiddlewareImplementation::class
 		)->getIterator();
 
-		self::assertEquals(
-			MiddlewareClassName::new( PassThroughMiddleware::class ),
+		self::assertSame(
+			PassThroughMiddleware::class,
 			$classNames->current()
 		);
 
 		$classNames->next();
 
-		self::assertEquals(
-			MiddlewareClassName::new( MiddlewareImplementation::class ),
+		self::assertSame(
+			MiddlewareImplementation::class,
 			$classNames->current()
 		);
 
@@ -119,9 +91,13 @@ final class MiddlewareClassNamesTest extends TestCase
 		self::assertFalse( $classNames->valid() );
 	}
 
+	/**
+	 * @throws ExpectationFailedException
+	 * @throws \Exception
+	 */
 	public function testAppend() : void
 	{
-		$classNames = MiddlewareClassNames::newFromStrings(
+		$classNames = MiddlewareClassNames::new(
 			PassThroughMiddleware::class,
 			MiddlewareImplementation::class
 		);
@@ -134,12 +110,12 @@ final class MiddlewareClassNamesTest extends TestCase
 		self::assertNotSame( $classNames, $appended );
 
 		$expectedMiddlewares = [
-			MiddlewareClassName::new( PassThroughMiddleware::class ),
-			MiddlewareClassName::new( MiddlewareImplementation::class ),
-			MiddlewareClassName::new( PassThroughMiddleware::class ),
-			MiddlewareClassName::new( MiddlewareImplementation::class ),
+			PassThroughMiddleware::class,
+			MiddlewareImplementation::class,
+			PassThroughMiddleware::class,
+			MiddlewareImplementation::class,
 		];
 
-		self::assertEquals( $expectedMiddlewares, iterator_to_array( $appended->getIterator(), false ) );
+		self::assertSame( $expectedMiddlewares, iterator_to_array( $appended->getIterator(), false ) );
 	}
 }