+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file. This project adheres
 to [Semantic Versioning](http://semver.org/) and [Keep a CHANGELOG](http://keepachangelog.com).
 
+## [3.0.0-beta3] - 2022-MM-DD
+
+* Make intialization with config optional
+* Allow direct setting of routes and app middlewares
+* Refactor middleware class names collection to a pure string collection
+
 ## [3.0.0-beta2] - 2022-05-27
 
 * Add ability to add middlewares to route groups, so middlewares have not to be repeated over multiple routes